pyproject.toml
[project]
name = "typstpresenter"
version = "0.1.0"
description = "Convert slides from PPTX to Typst (and maybe back?) using intermediate abstract representations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "python-pptx>=1.0.2",
    "typer>=0.24.1",
]

[project.scripts]
typstpresenter="typstpresenter.__main__:app"

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.8",
]
